Adding and deleting connection
> Allow user to add and delete connection
- When the user clicks the "+" button on the connections list page, the
user is taken to the add connection screen
- After entering the connection details, when the user clicks the
"save"button, the connection is added to the database.
- The connection is then retrieved from the database (including the
index) and added to the connections array as an object at that index.
- When the user clicks the "Edit" button on the connections list page,
a "-" sign appears beside each connection.
- Pressing this minus sign shows a delete button and the connection can
be deleted from the database and the array.
> Created a popup window for login
- When the user selects anyone of the connection, an alert window pops
up.
- If the user already gave the username and/or password while adding
the connection, the information is retrieved from the connection object
and displayed in the login window
- If not, the username and/or password field is left blank and the user
will need to enter this information before being allowed to proceed.
> Made some minor changes to the interface
- Deleted the Previous Login View Controller and its associated files.
- Changed the properties of the Add Connection View Controller.